mmi_backend/mmi_backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.3.0
python-dotenv==1.0.1
mysqlclient==2.2.4
WeasyPrint==62.3
celery==5.3.6
redis==5.0.3
django-filter==24.2
